cleanup

+++ db/migrate/20160409000553_add_consonants_to_dictionary_word.rb
@@ -1,8 +1,7 @@
 class AddConsonantsToDictionaryWord < ActiveRecord::Migration[5.0]
   def change
-    remove_column :dictionary_words, :hash_integer, :integer
-    add_column :dictionary_words, :hash_vowels, :string
-
-    add_index :dictionary_words, :hash_vowels
+    if column_exists? :dictionary_words, :hash_integer
+      remove_column :dictionary_words, :hash_integer, :integer
+    end
   end
 end
