+++ db/migrate/20160409060457_remove_hash_vowels_from_dictionary_word.rb
@@ -1,5 +1,7 @@
 class RemoveHashVowelsFromDictionaryWord < ActiveRecord::Migration[5.0]
   def change
-    remove_column :dictionary_words, :hash_vowels, :string
+    if column_exists? :dictionary_words, :hash_vowels
+      remove_column :dictionary_words, :hash_vowels, :string
+    end
   end
 end
